pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "puerml"
version = "0.1.18"
description = ""
authors = [ "Oleksii Zaichenko <oleksii@shipxdata.com>",]
readme = "README.md"

[tool.poetry.scripts]
test = "pytest:main"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pytest = "^8.1.1"
toml = "^0.10.2"
python-magic = "^0.4.27"
tqdm = "^4.66.2"
